service/requirements.txt
openai
supabase
feedparser
python-dotenv
